extracted search as component
